+++ changelog.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.18] - 2024-09-22
+## [0.1.20] - 2024-09-24
+### Added
+
+- pawn tests
+
+### Changed
+
+- fixed knight tests
+- fixed duplication on pawn methods
+
+## [0.1.19] - 2024-09-23
 ### Added
 
 - knight tests